+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.0.0-beta.1
-- Added support for `Packed` structs.
+- Added support for `Packed` structs. Packed annotations are generated
+automatically but can be overriden using `structs -> pack` config.
 - Updated sdk constraints to `>=2.13.0-211.6.beta`.
 
 # 2.4.2